+++ composeApp/src/commonMain/kotlin/tech/kotlinlang/permission/ui/component/UiComponentsScreen.kt
@@ -1,14 +1,21 @@
 package tech.kotlinlang.permission.ui.component
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.Button
 import androidx.compose.material.LocalTextStyle
+import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
@@ -18,6 +25,7 @@ import androidx.compose.ui.unit.sp
 import tech.kotlinlang.ui.AmountText
 import tech.kotlinlang.ui.AmountTextConfig
 import tech.kotlinlang.ui.AmountTextConfigLocal
+import kotlin.random.Random
 
 @Composable
 fun UiComponentsScreen() {
@@ -53,17 +61,27 @@ private fun AmountTextContent(modifier: Modifier) {
             decimalFontWeightDiff = 3,
         )
     ) {
-        Box(
+        Column(
             modifier = modifier,
-            contentAlignment = Alignment.Center,
+            horizontalAlignment = Alignment.CenterHorizontally,
+            verticalArrangement = Arrangement.spacedBy(12.dp),
         ) {
+            var amount by remember { mutableStateOf(Random.nextDouble(100.0, 1000.0)) }
             AmountText(
-                amount = 220.13,
+                amount = amount,
                 style = LocalTextStyle.current.copy(
                     fontWeight = FontWeight.Bold,
                     fontSize = 24.sp,
                 )
             )
+            Button(
+                onClick = {
+                    amount = Random.nextDouble(100.0, 1000.0);
+                },
+            ) {
+                Text("Random")
+            }
+
         }
     }
 }